+++ CHANGELOG.md
@@ -5,6 +5,15 @@ of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [17.21.18] - 2022-09-17
+
+### Added
+
+- Google reCAPTCHA v3 interceptor. Now you can easily add server side Google reCAPTCHA v3 validations, just add
+  the `common-clj.io.interceptors.auth/recaptcha-validation-interceptor` to the list of interceptors while handling
+  requests for http endpoints and add the `:recaptcha-secret-token` to the config file. Tested
+  with [reCAPTCHA v3](https://developers.google.com/recaptcha/docs/v3#programmatically_invoke_the_challenge).
+
 ## [17.20.18] - 2022-09-14
 
 ### Changed
@@ -337,7 +346,9 @@ of [keepachangelog.com](http://keepachangelog.com/).
 
 - Add `loose-schema` function.
 
-[Unreleased]: https://github.com/macielti/common-clj/compare/v17.20.18...HEAD
+[Unreleased]: https://github.com/macielti/common-clj/compare/v17.21.18...HEAD
+
+[17.21.18]: https://github.com/macielti/common-clj/compare/v17.20.18...v17.21.18
 
 [17.20.18]: https://github.com/macielti/common-clj/compare/v16.20.18...v17.20.18